test: Add visual tests
