+++ test/visualTests/features/Layout.test.ts
@@ -138,4 +138,25 @@ describe('LayoutTests', () => {
             settings
         );
     });
+
+    it('multibar-rests-single-track', async () => {
+        await VisualTestHelper.runVisualTest('layout/multibar-rest.gp', undefined, o => {
+            o.tracks = [0];
+            o.runs[0].referenceFileName = 'layout/multibar-rest-single-track.png';
+        });
+    });
+
+    it('multibar-rests-multi-track', async () => {
+        await VisualTestHelper.runVisualTest('layout/multibar-rest.gp', undefined, o => {
+            o.tracks = [0, 1];
+            o.runs[0].referenceFileName = 'layout/multibar-rest-multi-track.png';
+        });
+    });
+
+    it('multibar-rests-all-tracks', async () => {
+        await VisualTestHelper.runVisualTest('layout/multibar-rest.gp', undefined, o => {
+            o.tracks = [0, 1, 2];
+            o.runs[0].referenceFileName = 'layout/multibar-rest-all-tracks.png';
+        });
+    });
 });
